requirements.txt
# Core data processing
pandas
numpy

# Data modeling
pydantic

# Progress bars
tqdm

# Visualization
matplotlib
networkx
seaborn
plotly
kaleido

# Machine learning
scikit-learn
hdbscan

# LangChain and OpenAI API integration
langchain
langchain-core
langchain-community
langchain-openai
langchain-experimental
openai
azure-search-documents
azure-identity

# Configuration and utilities
toml
python-dotenv
Pillow

# Development tools
jupyter
notebook
streamlit